chore(Components): refactor react components to utilize styled components
* Update dashboard dependencieds
* fix testing-util versioning
* remove release-as in release-please config
* remove unused videojs css export on video-player
* add dashboard snapshot for UI CI task
