+++ packages/components/package.json
@@ -72,7 +72,7 @@
   "devDependencies": {
     "@cypress/vue": "^3.1.2",
     "@cypress/webpack-dev-server": "^1.8.4",
-    "@iot-app-kit/testing-util": "^3.0.0",
+    "@iot-app-kit/testing-util": "*",
     "@rollup/plugin-replace": "^3.0.0",
     "@stencil/router": "^1.0.1",
     "@types/chai": "^4.3.3",
