+++ packages/core/package.json
@@ -50,7 +50,7 @@
     "pack": "npm pack"
   },
   "dependencies": {
-    "@aws-sdk/client-iotsitewise": "^3.87.0",
+    "@aws-sdk/client-iotsitewise": "3.245.0",
     "d3-array": "^3.2.2",
     "intervals-fn": "^3.0.3",
     "parse-duration": "^1.0.3",
