+++ packages/dashboard/package.json
@@ -23,7 +23,7 @@
   "main": "dist/cjs/index.js",
   "module": "dist/esm/index.js",
   "scripts": {
-    "build": "npm run clean && rollup -c && npm run build:types",
+    "build": "npm run clean && rollup -c --bundleConfigAsCjs && npm run build:types",
     "build-storybook": "build-storybook",
     "build:types": "tsc --declarationDir dist/types --declaration true --emitDeclarationOnly true",
     "clean": "rimraf dist",
@@ -34,65 +34,64 @@
     "test:ui": "npx playwright test"
   },
   "devDependencies": {
-    "@aws-sdk/client-iotsitewise": "^3.245.0",
     "@iot-app-kit/jest-config": "^3.0.0",
     "@iot-app-kit/ts-config": "^3.0.0",
-    "@reduxjs/toolkit": "^1.8.6",
-    "@rollup/plugin-commonjs": "^22.0.2",
-    "@rollup/plugin-json": "^5.0.1",
-    "@rollup/plugin-node-resolve": "13.3.0",
-    "@storybook/addon-actions": "^6.5.12",
-    "@storybook/addon-essentials": "^6.5.12",
-    "@storybook/addon-interactions": "^6.5.12",
-    "@storybook/addon-links": "^6.5.12",
-    "@storybook/builder-webpack5": "^6.5.15",
-    "@storybook/manager-webpack5": "^6.5.15",
-    "@storybook/react": "^6.5.12",
+    "@reduxjs/toolkit": "^1.9.3",
+    "@rollup/plugin-commonjs": "^24.0.1",
+    "@rollup/plugin-json": "^6.0.0",
+    "@rollup/plugin-node-resolve": "15.0.1",
+    "@storybook/addon-actions": "^6.5.16",
+    "@storybook/addon-essentials": "^6.5.16",
+    "@storybook/addon-interactions": "^6.5.16",
+    "@storybook/addon-links": "^6.5.16",
+    "@storybook/builder-webpack5": "^6.5.16",
+    "@storybook/manager-webpack5": "^6.5.16",
+    "@storybook/react": "^6.5.16",
     "@storybook/testing-library": "^0.0.13",
-    "@testing-library/dom": "^8.20.0",
+    "@testing-library/dom": "^9.0.1",
     "@testing-library/jest-dom": "^5.16.5",
     "@testing-library/react": "^14.0.0",
     "@testing-library/user-event": "^14.4.3",
     "@types/box-intersect": "^1.0.0",
     "@types/is-hotkey": "^0.1.7",
     "@types/lodash": "^4.14.191",
-    "@types/node": "^18.11.4",
+    "@types/node": "^18.15.3",
     "@types/react": "^18.0.28",
     "@types/react-dom": "^18.0.11",
-    "css-loader": "5.2.6",
+    "css-loader": "6.7.3",
     "dotenv": "^16.0.3",
     "eslint-config-iot-app-kit": "3.0.0",
     "jest": "^29.5.0",
     "jest-environment-jsdom": "^29.5.0",
     "jest-extended": "^3.2.4",
-    "jest-matcher-utils": "^29.3.1",
+    "jest-matcher-utils": "^29.5.0",
     "lodash": "^4.17.21",
-    "postcss": "^8.2.4",
-    "postcss-import": "^14.0.0",
+    "postcss": "^8.4.21",
+    "postcss-import": "^15.1.0",
     "postcss-url": "^10.1.3",
     "react": "^18.2.0",
     "react-dnd-test-backend": "^16.0.1",
     "react-dnd-test-utils": "^16.0.1",
     "react-dom": "^18.2.0",
-    "react-redux": "^8.0.4",
-    "redux": "^4.2.0",
-    "rimraf": "^3.0.2",
-    "rollup": "2.76.0",
+    "react-redux": "^8.0.5",
+    "redux": "^4.2.1",
+    "rimraf": "^4.4.0",
+    "rollup": "3.19.1",
     "rollup-plugin-peer-deps-external": "^2.2.4",
     "rollup-plugin-postcss": "^4.0.2",
     "rollup-plugin-tsconfig-paths": "^1.4.0",
     "rollup-plugin-typescript2": "^0.34.1",
-    "style-loader": "2.0.0",
-    "tsconfig-paths-webpack-plugin": "^4.0.0",
-    "tslib": "^2.4.0",
+    "style-loader": "3.3.2",
+    "tsconfig-paths-webpack-plugin": "^4.0.1",
+    "tslib": "^2.5.0",
     "typescript": "^4.9.5",
-    "webpack": "^5.76.1"
+    "webpack": "^5.76.2"
   },
   "dependencies": {
-    "@aws-sdk/client-iotsitewise": "^3.245.0",
-    "@cloudscape-design/components": "^3.0.227",
-    "@cloudscape-design/design-tokens": "^3.0.7",
-    "@cloudscape-design/global-styles": "^1.0.5",
+    "@aws-sdk/client-iotsitewise": "3.245.0",
+    "@cloudscape-design/components": "^3.0.228",
+    "@cloudscape-design/design-tokens": "^3.0.9",
+    "@cloudscape-design/global-styles": "^1.0.7",
     "@iot-app-kit/components": "^4.0.0",
     "@iot-app-kit/core": "^4.0.0",
     "@iot-app-kit/react-components": "4.0.0",
