+++ packages/dashboard/src/components/resizablePanes/index.tsx
@@ -6,6 +6,7 @@ export const LEFT_WIDTH_PERCENT = 0.2;
 export const RIGHT_WIDTH_PERCENT = 0.2;
 export const MINIMUM_CENTER_PANE_WIDTH = 100;
 export const MINIMUM_SIDE_PANE_WIDTH = 100;
+const DEFAULT_SIDE_PANE_WIDTH = 250;
 export const HANDLE_WIDTH = 10;
 export const MAXIMUM_PANES_PROPORTION = 0.8;
 export const LEFT_WIDTH_PERCENT_STORAGE_KEY = 'iot-dashboard-pane-left-width';
@@ -39,8 +40,8 @@ export const ResizablePanes: FC<ResizablePanesProps> = ({ leftPane, centerPane, 
   const [movedX, setMovedX] = useState<number | null>(null);
 
   /** Current widths of the three panes, in px */
-  const [leftPaneWidth, setLeftPaneWidth] = useState(MINIMUM_SIDE_PANE_WIDTH);
-  const [rightPaneWidth, setRightPaneWidth] = useState(MINIMUM_SIDE_PANE_WIDTH);
+  const [leftPaneWidth, setLeftPaneWidth] = useState(DEFAULT_SIDE_PANE_WIDTH);
+  const [rightPaneWidth, setRightPaneWidth] = useState(DEFAULT_SIDE_PANE_WIDTH);
 
   useEffect(() => {
     // On initial load, attempt to get stored widths from sessionStorage.
