+++ packages/react-components/e2e/tests/status/status.spec.ts
@@ -2,12 +2,13 @@ import { test, expect } from '@playwright/test';
 
 const TEST_PAGE = 'http://localhost:6007/?path=/story/widgets-status-status-base--main';
 const TEST_IFRAME = '#storybook-preview-iframe';
-const COMPONENT_SELECTOR = '.status-widget';
+const COMPONENT_SELECTOR = '[data-testid="status-widget"]';
 
 test('status', async ({ page }) => {
   await page.goto(TEST_PAGE);
   const frame = page.frameLocator(TEST_IFRAME); // Need to go into frame otherwise the `locator` won't locate the selection.
 
+  await expect(frame.locator(COMPONENT_SELECTOR)).toContainText("Windmill");
   await expect(frame.locator(COMPONENT_SELECTOR)).toHaveScreenshot('default.png');
 
   // unit will display
