+++ packages/react-components/jest.config.ts
@@ -3,6 +3,14 @@ import { reactConfig } from '@iot-app-kit/jest-config';
 const config = {
   ...reactConfig,
   roots: ['src'],
+  coverageThreshold: {
+    global: {
+      statements: 80,
+      branches: 70,
+      functions: 70,
+      lines: 80,
+    },
+  },
 };
 
 export default config;
