+++ packages/react-components/package.json
@@ -15,43 +15,44 @@
   },
   "version": "4.0.0",
   "scripts": {
+    "start": "npx start-storybook -p 6007",
+    "dev": "npm-watch build",
+    "clean": "rm -rf dist && rm -rf storybook-static && rm -rf test-results && rm -rf coverage",
     "build": "npm run clean && npm run build:es && npm run build:cjs",
     "build:es": "tsc -p ./tsconfig.build.json --outDir ./dist/es",
     "build:cjs": "tsc -p ./tsconfig.build.json --outDir ./dist/cjs --module CommonJS",
-    "clean": "rm -rf dist && rm -rf storybook-static",
-    "dev": "npm-watch build",
-    "copy:license": "cp ../../LICENSE LICENSE",
-    "copy:notice": "cp ../../NOTICE NOTICE",
-    "prepack": "npm run copy:license && npm run copy:notice",
-    "pack": "npm pack",
     "test": "npm run test:jest && npm run test:types",
     "test:jest": "TZ=utc jest",
     "test:types": "tsc -p ./tsconfig.json --noEmit",
     "test:watch": "TZ=utc jest --watch",
-    "lint": "eslint . --max-warnings=0",
-    "fix": "eslint --fix .",
     "test:ui": "npx playwright test",
     "test:ui-update": "npx playwright test --update-snapshots",
-    "start": "npx start-storybook -p 6007"
+    "lint": "eslint . --max-warnings=0",
+    "fix": "eslint --fix .",
+    "copy:license": "cp ../../LICENSE LICENSE",
+    "copy:notice": "cp ../../NOTICE NOTICE",
+    "prepack": "npm run copy:license && npm run copy:notice",
+    "pack": "npm pack"
   },
   "watch": {
     "build": {
       "patterns": [
         "src"
       ],
-      "extensions": "ts,html,css,tsx",
+      "extensions": "ts,html,tsx",
       "quiet": "false"
     }
   },
   "devDependencies": {
     "@iot-app-kit/jest-config": "*",
     "@iot-app-kit/source-iotsitewise": "^4.0.0",
-    "@iot-app-kit/testing-util": "^3.0.0",
+    "@iot-app-kit/testing-util": "*",
     "@iot-app-kit/ts-config": "*",
     "@jest/globals": "^29.5.0",
     "@storybook/addon-essentials": "^6.5.16",
     "@storybook/addon-links": "^6.5.16",
     "@storybook/addons": "^6.5.16",
+    "@storybook/builder-webpack5": "^6.5.16",
     "@storybook/preset-scss": "^1.0.3",
     "@storybook/react": "^6.5.16",
     "@storybook/testing-library": "^0.0.13",
@@ -59,10 +60,11 @@
     "@types/color": "^3.0.3",
     "@types/d3": "^7.4.0",
     "@types/dompurify": "2.3.3",
-    "@types/jest": "^29.4.0",
+    "@types/jest": "^29.4.4",
     "@types/lodash.omitby": "^4.6.7",
     "@types/react": "^18.0.28",
     "@types/react-dom": "^18.0.11",
+    "@types/styled-components": "^5.1.26",
     "@types/uuid": "^9.0.1",
     "@types/video.js": "^7.3.15",
     "babel-plugin-formatjs": "10.3.18",
@@ -78,7 +80,7 @@
   },
   "dependencies": {
     "@cloudscape-design/collection-hooks": "^1.0.19",
-    "@cloudscape-design/components": "^3.0.226",
+    "@cloudscape-design/components": "^3.0.228",
     "@iot-app-kit/components": "4.0.0",
     "@iot-app-kit/core": "4.0.0",
     "@iot-app-kit/source-iottwinmaker": "4.0.0",
@@ -88,6 +90,7 @@
     "dompurify": "2.3.4",
     "lodash.omitby": "^4.6.0",
     "parse-duration": "^1.0.3",
+    "styled-components": "^5.3.9",
     "uuid": "^9.0.0",
     "video.js": "7.20.3"
   },
