--- packages/react-components/src/components/shared-components/ErrorBadge/ErrorBadge.css
@@ -1,13 +1,0 @@
-.error-badge div {
-  display: flex;
-  padding-left: 1.4286em;
-  line-height: 1.4286em;
-  color: var(--error-badge-font-color, #6a7070);
-}
-
-.error-badge .warning-symbol {
-  font-size: 2rem;
-  padding-right: var(--margin-small);
-  font-weight: bold;
-  color: var(--error-badge-warning-color, #fa3232);
-}
