--- packages/react-components/src/components/shared-components/LoadingSpinner/LoadingSpinner.css
@@ -1,31 +1,0 @@
-@keyframes fade-in {
-  from {
-    opacity: 0;
-  }
-
-  to {
-    opacity: 1;
-  }
-}
-
-@keyframes spin {
-  from {
-    transform: rotate(0);
-  }
-
-  to {
-    transform: rotate(360deg);
-  }
-}
-
-.loading-spinner {
-  animation: spin 900ms linear infinite, fade-in 1s linear;
-}
-
-path#c {
-  fill: var(--loading-spinner-color);
-}
-
-.dark path#c {
-  fill: var(--loading-spinner-color-dark);
-}
