--- packages/react-components/src/components/shared-components/StatusIcon/StatusIcon.css
@@ -1,6 +1,0 @@
-.status-icon {
-  position: relative;
-  margin-right: 3px;
-  top: 2px;
-  display: inline;
-}
