+++ packages/react-components/src/components/status/statusBase.tsx
@@ -1,12 +1,43 @@
 import React from 'react';
-import './status.css';
 import { DEFAULT_STATUS_SETTINGS, DEFAULT_STATUS_COLOR, STATUS_ICON_SHRINK_FACTOR } from './constants';
 
 import { ErrorBadge, LoadingSpinner, StatusIcon, Value } from '../shared-components';
 import { highContrastColor } from './highContrastColor';
 import { DEFAULT_MESSAGE_OVERRIDES } from '../../common/dataTypes';
 import omitBy from 'lodash.omitby';
+import styled from 'styled-components';
 import type { StatusProperties, StatusSettings } from './types';
+
+const StatusWidget = styled.div`
+  display: flex;
+  height: 100%;
+  box-sizing: border-box;
+  flex-direction: column;
+  padding: var(--margin-medium);
+  font-size: var(--font-size-3);
+  line-height: var(--line-height-3);
+`;
+
+const Secondary = styled.div`
+  font-size: var(--font-size-1);
+  line-height: var(--line-height-1);
+  font-weight: var(--font-weight-light);
+`;
+
+const ValueContainer = styled.div`
+  font-weight: var(--font-weight-bold);
+  font-size: var(--font-size-3);
+  line-height: var(--line-height-3);
+`;
+
+const Spacer = styled.div`
+  width: 4px;
+  display: inline-block;
+`;
+
+const Divider = styled.div`
+  flex-grow: 1;
+`;
 
 export const StatusBase: React.FC<StatusProperties> = ({
   icon,
@@ -32,36 +63,32 @@ export const StatusBase: React.FC<StatusProperties> = ({
 
   /** Display Settings. We want to dynamically construct the layout dependent on what information is shown */
   const emphasizeValue = !showValue;
-  const emphasizeNameAndUnit = showValue && !showName && !showUnit;
 
   /** If anything is emphasized, then something is emphasized */
-  const somethingIsEmphasized = emphasizeValue || emphasizeNameAndUnit;
 
   return (
-    <div
-      className='status-widget'
+    <StatusWidget
+      data-testid="status-widget"
       style={{
         backgroundColor,
         color: foregroundColor,
-        justifyContent: somethingIsEmphasized ? 'center' : 'unset',
+        justifyContent: emphasizeValue ? 'center' : 'unset',
       }}
     >
       {showName && name}
       {error && <ErrorBadge>{error}</ErrorBadge>}
       {isLoading && <LoadingSpinner size={fontSize} />}
 
-      {!somethingIsEmphasized && !isLoading && <div className='divider' />}
+      {!emphasizeValue && !isLoading && <Divider />}
       {showValue && point && !isLoading && (
-        <div className={emphasizeNameAndUnit ? 'center' : ''}>
+        <div>
           {alarmPoint && propertyPoint && (
-            <div className='secondary'>
-              <span style={{ color: foregroundColor }}>
-                {DEFAULT_MESSAGE_OVERRIDES.liveTimeFrameValueLabel}:{' '}
-                <Value value={propertyPoint ? propertyPoint.y : undefined} unit={showUnit ? unit : undefined} />
-              </span>
-            </div>
+            <Secondary style={{ color: foregroundColor }}>
+              {DEFAULT_MESSAGE_OVERRIDES.liveTimeFrameValueLabel}:{' '}
+              <Value value={propertyPoint ? propertyPoint.y : undefined} unit={showUnit ? unit : undefined} />
+            </Secondary>
           )}
-          <div className={`value ${emphasizeNameAndUnit ? 'large' : ''}`} style={{ color: foregroundColor }}>
+          <div style={{ color: foregroundColor, display: 'flex' }}>
             {showIcon && icon && (
               <>
                 <StatusIcon
@@ -69,13 +96,15 @@ export const StatusBase: React.FC<StatusProperties> = ({
                   size={fontSize * STATUS_ICON_SHRINK_FACTOR}
                   color={highContrastColor(backgroundColor)}
                 />
-                <div className='spacer' />
+                <Spacer />
               </>
             )}
-            <Value value={point ? point.y : undefined} unit={showUnit && alarmPoint == null ? unit : undefined} />
+            <ValueContainer>
+              <Value value={point ? point.y : undefined} unit={showUnit && alarmPoint == null ? unit : undefined} />
+            </ValueContainer>
           </div>
         </div>
       )}
-    </div>
+    </StatusWidget>
   );
 };
