+++ packages/react-components/src/components/video-player/requestVideoUpload.tsx
@@ -1,6 +1,5 @@
 import { Button, DateRangePicker, SpaceBetween } from '@cloudscape-design/components';
 import React from 'react';
-import 'video.js/dist/video-js.css';
 import { i18nStrings, requestVideoButtonLabel } from './constants';
 import { getStartAndEndTimeFromRange } from './utils/dateRangeUtils';
 import type { DateRangePickerProps } from '@cloudscape-design/components';
