+++ packages/react-components/tsconfig.build.json
@@ -4,6 +4,7 @@
     "./src"
   ],
   "exclude": [
-    "**/*.spec.*"
+    "**/*.spec.*",
+    "**/*.css"
   ]
 }
