+++ packages/source-iottwinmaker/package.json
@@ -50,7 +50,7 @@
     "pack": "npm pack"
   },
   "dependencies": {
-    "@aws-sdk/client-iotsitewise": "^3.131.0",
+    "@aws-sdk/client-iotsitewise": "3.245.0",
     "@aws-sdk/client-iottwinmaker": "^3.131.0",
     "@aws-sdk/client-kinesis-video": "^3.131.0",
     "@aws-sdk/client-kinesis-video-archived-media": "^3.131.0",
